Changes to "include" variables, style, and phase vs time plot
Changed both include_column and include_vds to vds.include (variable
for the include column number)…not quite sure if they’re the same thing
but the script still works.

A few stylistic changes.

Started the phase vs time plot (raw data should be plotted and saved as
a graph).